Completed The Admin Interface: 1-Fixed StatsController, also added View for it. 2-Account Controller, Login Controller, Request Management Controller: dealt with nullable variables. 3-Search Controller: Adjusted Minor things, spaces and such. 4-Index Home: now shows the actual stats according to roles. 5-added Request Scholarship method to Login Controller, also created the view for it. 6-Search/View: Fixed Search button. 7-Nothing on _Layout 8-Added classes for progress bar in site.css 9-updated packages.

+++ Triplite Committee Platform/Controllers/RequestManagementController.cs
@@ -176,6 +176,11 @@ namespace Triplite_Committee_Platform.Controllers
         {
             var requestTypeModel = await _context.RequestType.FindAsync(id);
             var existingReasons = await _context.Reasons.Where(d => d.ReqTypeID == id).FirstOrDefaultAsync();
+            if(requestTypeModel == null)
+            {
+                TempData["Error"] = "Request Type Not Found.";
+                return RedirectToAction("Index");
+            }
             if (existingReasons != null)
             {
                 _context.Reasons.Remove(existingReasons);
